### 2024-05-26 #6784 @Kevin-Mattheus-Moerman
@Peter-Hogg thanks for this submission. As AEiC on this track I am currently determining if this work is in scope for JOSS. Due to its relatively small size and limited functionality offered I have just added the `query-scope` label, which will trigger a scope review by the editorial board.

Below are some additional findings by myself after having a quick look at  your repository.

- [ ] The README makes reference to `pynt`, is that the old name of the software?
- [ ] The following URLs in the README are not functional:
 `....please consult the [README](https://github.com/padster/pyNeuroTrace/tree/master/pyneurotrace/pyneurotrace) next to viz.py.`
`...filters are provided within [filters.py](https://github.com/padster/pyNeuroTrace/blob/master/pyneurotrace/pyneurotrace/filters.py).`
`...are available within [files.py](https://github.com/padster/pyNeuroTrace/blob/master/pyneurotrace/pyneurotrace/files.py), including:`
`... from experiments performed by the [Kurt Haas lab](http://www.haaslab.com/)`
- [ ] From looking at the repository it appears that documentation and automated testing are missing. If so this is not acceptable for JOSS.
- [ ] The software appears to lack contributing guidelines

I'm afraid that as it stands this work is not in scope for JOSS given these points.
