+++ docs/docs.md
@@ -1,0 +1,1065 @@
+# pyNeuroTrace Documentation
+
+The will file will provide and overview documentation of the pyNeuroTrace library
+
+
+# Table of Contents
+1. [Installation](#Installation)
+2. [Overview](#Overview)
+3. [Modules](#Modules)
+    - [events](#events)
+    - [files](#files)
+    - [filters](#filters)
+    - [morphology](#morphology)
+    - [notebook](#notebook)
+    - [viz](#viz)
+4. [Examples][#Examples]        
+
+
+# Installation
+`pyNeuroTrace` can be installed with pip:
+
+```
+pip install pyNeuroTrace
+```
+
+If you wish to use optional GPU accelrated modules install `pyNeuroTrace` with this modification to the above command:
+
+```
+pip install pyNeuroTrace['GPU']
+```
+Or insure the `Cupy` is installed in your Python environment and that you have a CUDA enabled graphics card.
+
+# Overview
+This is an overview of the useful modules in the library
+
+```plaintext
+pyneurotrace
+├── gpu
+│   ├── events
+│   ├── filters
+├── events
+├── files
+├── filters
+├── morphology
+├── notebook
+├── viz
+```
+# Modules
+Here is an overview of the core modules and their functions
+## events
+Documentation for the functions in the events module
+### ewma
+
+`pyneurotrace.events.ewma(data, weight=0.1)`
+
+Performs smoothing on each row using the Exponentially Weighted Moving Average (EWMA) method.
+
+#### Parameters:
+
+- `data`: array
+  - Data array to be smoothed.
+- `weight`: float, optional
+  - Weight for new values versus existing average. Default is 0.1.
+
+#### Returns:
+
+- `smoothed_data`: array
+  - Smoothed data array.
+___
+
+### cusum
+
+`pyneurotrace.events.cusum(data, slack=1.0)`
+
+Calculates the Cumulative Sum (CUSUM) of movement above the mean for each row in the data.
+
+#### Parameters:
+
+- `data`: array
+  - Data array to be analyzed.
+- `slack`: float, optional
+  - Slack noise value. Default is 1.0.
+
+#### Returns:
+
+- `cusum_data`: array
+  - CUSUM data array.
+___
+
+### matchedFilter
+
+`pyneurotrace.events.matchedFilter(data, hz, windowSize, A=2.0, riseRate=0.028, decayRate=0.39)`
+
+Performs a likelihood ratio test using a Matched Filter (MF) to see how well a previous window matches a desired signal over no signal.
+
+#### Parameters:
+
+- `data`: array
+  - Data array to be filtered.
+- `hz`: int
+  - Sampling rate in Hz.
+- `windowSize`: int
+  - Size of the window to match.
+- `A`: float, optional
+  - Amplitude of the desired signal. Default is 2.0.
+- `riseRate`: float, optional
+  - Rise rate of the desired signal. Default is 0.028.
+- `decayRate`: float, optional
+  - Decay rate of the desired signal. Default is 0.39.
+
+#### Returns:
+
+- `mf_data`: array
+  - Matched filter data array.
+___
+
+### thresholdEvents
+
+`pyneurotrace.events.thresholdEvents(data, threshold, minBelowBefore=1)`
+
+Turns an event detector into event yes/no indicators by applying a threshold and only keeping events that occur after a minimum number of non-event samples.
+
+#### Parameters:
+
+- `data`: array
+  - Data array containing event detector outputs.
+- `threshold`: float
+  - Confidence threshold for events.
[…]
