+++ CONTRIBUTING.md
@@ -1,0 +1,44 @@
+# Introduction
+
+### Contributing to PyNeuroTrace!
+
+Thank you for considering contributing to PyNeuroTrace. We're alway happy to work with more people in the development of our tool.
+
+
+### Types of contributions to PyNeuroTrace
+
+We are open to new features for either plotting or proccesing time-series data. We also welcome bug reports and ask you follow the guidelines below if you discover a bug and wish to report it. 
+
+
+# Getting started
+### How to suggest a feature or enhancement
+If you have a filter that you want us to add to pyNeuroTrace open an issue and include a citation to the filter. We will review proposal and maybe able to add it to the library.
+
+### Contributing your own code
+Please first open an issue on the repository outlining the scope of what you want to add for us to review before you start work. Then do the following once the feature is approved:
+
+>1. Create your own fork of pyNeuroTrace
+>2. Add the feature and test it on the fork
+>3. Once complete open a pull request
+
+We will try to review all pull-requests in a timely fashion.
+
+### Small changes.
+
+For small bugs or changes please open an issue see below
+
+# Filing a bug report.
+Open an issue on the on [GitHub repository](https://github.com/padster/pyNeuroTrace/issues) use the following template
+
+
+> 1. What version of Python are you using? What version are the relevant libraries? 
+> 2. What operating system and processor architecture are you using?
+> 3. What did you do?
+> 4. What did you expect to see?
+> 5. What did you see instead?
+> 6. What was the nature of the data you were working with? (type and dimensions)
+
+
+
+### Acknowledgments
+This contributing docuement was based off of [@nayafia](https://github.com/nayafia)'s [contributing template](https://github.com/nayafia/contributing-template) 
