+++ requirements.txt
@@ -3,6 +3,7 @@ ipywidgets
 matplotlib
 matplotlib-scalebar
 numpy
+pytest
 PyQt5
 scipy
 tqdm
