+++ .github/workflows/python-test.yml
@@ -19,7 +19,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-
+        # Install optional dependencies
+        pip install cupy
     - name: Run tests
       run: |
         pytest
