+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+# Ignore specific directories or files
+norecursedirs = .git __pycache__ node_modules pyneurotrace/gpu
