+++ pyneurotrace/__init__.py
@@ -1,4 +1,11 @@
-from . import gpu
+try:
+    import cupy as cp
+    from . import gpu
+    has_cupy = True
+except ImportError:
+    has_cupy = False
+    gpu = None
+
 from . import events
 from . import filters
 from . import files
