### 2024-05-27 f7ec4e9
Added tests for events
