### 2024-05-27 #6784 @Kevin-Mattheus-Moerman
Thanks @Peter-Hogg can you comment on documentation and testing. Do you have: 1) dedicated documentation, 2) automated testing?

Here are some example python projects which have the documentation (and often a dedicated website for it) as well as testing approaches clear shown:

https://github.com/sunpy/sunpy
https://github.com/scipy/scipy

and for instance this package (which is currently under review in JOSS too): https://github.com/FixOurFood/AgriFoodPy/, which has a docs and testing badges clearly at the top.

I recommend that you similarly have such badges or at least that the README section has a documentation and a testing section.
