### 2024-05-28 #6784 @Kevin-Mattheus-Moerman
@Peter-Hogg this submission has passed the initial scope check by the editorial board. However all agree this submission will require additional work before being acceptable. I've hinted at documentation and testing but the handling editor and reviewer will raise more points to work on. Please ensure you can implement such changes in a timely fashion.
