+++ docs/source/installation.rst
@@ -1,0 +1,17 @@
+Installation
+============
+
+pyNeuroTrace can be installed with pip:
+
+.. code-block:: shell
+
+   pip install pyNeuroTrace
+
+If you wish to use optional GPU-accelerated modules, install pyNeuroTrace with this modification:
+
+.. code-block:: shell
+
+   pip install pyNeuroTrace[GPU]
+
+Or ensure that `Cupy` is installed in your Python environment and that you have a CUDA-enabled graphics card.
+

[…]
