+++ test/test_viz.py
@@ -30,10 +30,22 @@ def test_plotLine():
     plt.close(fig)
 
 def test_plotAveragePostStimIntensity():
-    data = np.random.random((10, 1000))
+    def generate_exponential_decay_signal(length, tau):
+        t = np.arange(length)
+        signal = np.exp(-t / tau) + 0.1 * np.random.normal(size=length)
+        return signal
+
+    # Parameters
+    num_signals = 10
+    signal_length = 1000
+    tau = 20  # Time constant for the decay
     hz = 10
     stimOffIdx = np.array([20, 30, 40, 50, 60, 70, 80, 90, 100])
     stimOnIdx = np.array([25, 35, 45, 55, 65, 75, 85, 95, 105])
+
+
+    data = np.array([generate_exponential_decay_signal(signal_length, tau) for _ in range(num_signals)])
+   
     fig, ax = plt.subplots()
     
     # Run the function to ensure it generates the plot without errors
