### 2024-05-29 #6784 @Peter-Hogg
> Hi @Peter-Hogg, I'll take over the submission from now on.
>
> Would you be able to recommend appropriate reviewers for this submission? You can use the [list of potential reviewers](https://reviewers.joss.theoj.org/reviewers) to help. Please provide a list, but don't tag them (with @) so they aren't notified unnecessarily.
>
> Thanks!

I looked through the list of reviewers and the following seems like good fits for this work:
Yannan Chen
Daniel Müller-Komorowska
suhang96
edeno
