### 2024-05-29 3773307
Docs workflow
Adds a workflow yaml for docs generation
