+++ .github/workflows/sphinx.yml
@@ -1,43 +1,36 @@
 name: Sphinx Build
 on:
-  [push]
+  push:
+    branches:
+      - main  # or the branch you want to trigger the action
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v1
-    # Standard drop-in approach that should work for most people.
-    - uses: ammaraskar/sphinx-action@master
+    - name: Checkout repository
+      uses: actions/checkout@v2
+
+    - name: Set up Python
+      uses: actions/setup-python@v2
       with:
-        build-command: "sphinx-build -b html . _build"
-        docs-folder: "docs/"
-    # Great extra actions to compose with:
-    # Create an artifact of the html output.
-    - uses: actions/upload-artifact@v1
-      with:
-        name: DocumentationHTML
-        path: docs/_build/html/
-    # Publish built docs to gh-pages branch.
-    # ===============================
-    - name: Commit documentation changes
+        python-version: '3.x'  # Specify the Python version
+
+    - name: Install dependencies
       run: |
-        git clone https://github.com/ammaraskar/sphinx-action-test.git --branch gh-pages --single-branch gh-pages
-        cp -r docs/_build/html/* gh-pages/
-        cd gh-pages
-        touch .nojekyll
-        git config --local user.email "action@github.com"
-        git config --local user.name "GitHub Action"
-        git add .
-        git commit -m "Update documentation" -a || true
-        # The above command will fail if no changes were present, so we ignore
-        # that.
-    - name: Push changes
-      uses: ad-m/github-push-action@master
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+        pip install sphinx sphinx_rtd_theme
+
+    - name: Build documentation
+      run: |
+        cd docs  # Change to your documentation directory
+        make html
+
+    - name: Deploy to GitHub Pages
+      uses: peaceiris/actions-gh-pages@v3
       with:
-        branch: gh-pages
-        directory: gh-pages
         github_token: ${{ secrets.GITHUB_TOKEN }}
-    # ===============================
+        publish_dir: ./docs/_build/html  # Path to the built HTML files
+        publish_branch: gh-pages
