+++ .github/workflows/sphinx.yml
@@ -22,6 +22,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install -r requirements.txt
         pip install sphinx sphinx_rtd_theme
+        pip install myst-parser
 
     - name: Build documentation
       run: |
