### 2024-05-29 8f14227
Update test_viz.py
