+++ .github/workflows/sphinx.yml
@@ -33,5 +33,5 @@ jobs:
       uses: peaceiris/actions-gh-pages@v3
       with:
         github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: ./docs/_build/html  # Path to the built HTML files
+        publish_dir: ./docs/html  # Path to the built HTML files
         publish_branch: gh-pages
