### 2024-05-29 #6784 @adamltyson
Thanks for the reviewer suggestions @Peter-Hogg!

@danielmk, @suhang96, @edeno,  @Yannan-Chen would you be able to review this submission to JOSS?

If you haven't reviewed for JOSS before, we carry out our checklist-driven reviews here in GitHub issues following [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). If you have any questions, let me know.
