+++ docs/build/html/_sources/modules/index.rst.txt
@@ -1,0 +1,14 @@
+Modules
+=======
+
+Here is an overview of the core modules and their functions.
+
+.. toctree::
+   :maxdepth: 1
+
+   events
+   files
+   filters
+   morphology
+   notebook
+   viz
