+++ docs/build/html/_sources/modules/viz.rst.txt
@@ -1,0 +1,14 @@
+Visualization
+=============
+
+Tools for visualizing experimental data.
+
+.. autofunction:: pyneurotrace.viz.plotIntensity
+.. autofunction:: pyneurotrace.viz.plotLine
+.. autofunction:: pyneurotrace.viz.plotAveragePostStimIntensity
+.. autofunction:: pyneurotrace.viz.plotAveragePostStimTransientParams
+.. autofunction:: pyneurotrace.viz.plotPlanarStructure
+.. autofunction:: pyneurotrace.viz.plotBaseTipScatter
+.. autofunction:: pyneurotrace.viz.planarAnimation
+.. autofunction:: pyneurotrace.viz.kymograph
+.. autofunction:: pyneurotrace.viz.scrollingAnimation
