+++ docs/build/html/_sources/index.rst.txt
@@ -1,0 +1,23 @@
+.. pyNeuroTrace documentation master file, created by
+   sphinx-quickstart on Thu May 27 2024.
+
+Welcome to pyNeuroTrace's documentation!
+========================================
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Table of Contents
+
+   installation
+   overview
+   modules/index
+   examples
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
