+++ docs/build/html/_sources/modules/morphology.rst.txt
@@ -1,0 +1,8 @@
+Morphology
+==========
+
+These functions are used to reconstruct a neuronal structure imaged using a custom AOD microscope.
+
+.. autofunction:: pyneurotrace.morphology.treePostProcessing
+.. autofunction:: pyneurotrace.morphology.buildBranchIDMap
+.. autofunction:: pyneurotrace.morphology.treeToFiloTipAndBase
