+++ docs/build/html/_sources/examples.rst.txt
@@ -1,0 +1,4 @@
+Examples
+========
+
+Please look at the Jupyter notebooks in the GitHub repository for examples.
