+++ .github/workflows/python-test.yml
@@ -1,4 +1,4 @@
-name: Python package
+name: PyTests
 
 on: [push, pull_request]
 
