+++ .github/workflows/sphinx.yml
@@ -1,4 +1,4 @@
-name: Sphinx Build
+name: Documentation
 on:
   push:
     branches:
