### 2024-05-30 #6784 @danielmk
Dear Adam, thank you for the invitation. I'd be happy to review
PyNeuroTrace.

Best,
Daniel
