### 2024-06-11 #6877 @editorialbot
[REVIEW]: PyNeuroTrace - Python code for neural activity time series
**Submitting author:** @peter-hogg (<a href="http://orcid.org/0000-0003-2176-4977">Peter William Hogg</a>)
**Repository:** https://github.com/padster/pyNeuroTrace
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.1
