### 2024-06-13 padster/pyNeuroTrace#4 @danielmk
error importing pyneurotrace after installation
This issue is part of the review at openjournals/joss-reviews#6877

This issue concerns the installation instructions.

I created a new environment with `conda create --name pyneurotrace python=3.9` and then in the new environment ran `pip install pyneurotrace` as per the installation instructions. pip does not raise any errors. Then in Python I get the following error trying to `import pyneurotrace`:
```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "C:\Users\Daniel\anaconda3\envs\pyneurotrace\lib\site-packages\pyneurotrace\__init__.py", line 1, in <module>
    from . import gpu
  File "C:\Users\Daniel\anaconda3\envs\pyneurotrace\lib\site-packages\pyneurotrace\gpu\__init__.py", line 1, in <module>
    from . import filters
  File "C:\Users\Daniel\anaconda3\envs\pyneurotrace\lib\site-packages\pyneurotrace\gpu\filters.py", line 2, in <module>
    import cupy as cu
ModuleNotFoundError: No module named 'cupy'
```

Did I miss something? I'm on windows 10.
