### 2024-06-13 padster/pyNeuroTrace#4 @Peter-Hogg
Hi Daniel,

Something is not working as it should.  There appears to be an issue with the optional dependency Cupy. I thought it was setup so that  the base install would ignore the GPU versions. I'll investigate this and give you an update once I resolve this issue.
