+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyNeuroTrace"
-version = "0.9.5"
+version = "0.9.6"
 dependencies = [
     "numpy",
     "scipy",
