+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyNeuroTrace"
-version = "0.9.6"
+version = "0.9.7"
 dependencies = [
     "numpy",
     "scipy",
