### 2024-06-13 padster/pyNeuroTrace#4 @Peter-Hogg
I believe I have fixed the issue
