### 2024-06-18 padster/pyNeuroTrace#4 @danielmk
Thank you for the fix. I can indeed import pyneurotrace now. Without having cupy installed. If I understand correctly, the pip install described in the readme installs pyneurotrace always in a way that it does not support the GPU implementation. I think that's generally good, since users should be able to explore the package without worrying about GPU. However, since the GPU implementation is a major part of the JOSS submission, I'd suggest that the readme should explain how to install with GPU support.
