### 2024-06-18 padster/pyNeuroTrace#5 @danielmk
OSError on pyntViz.plotIntensity
This issue is part of the review at https://github.com/openjournals/joss-reviews/issues/6877

I've tried going through the notebook. In the second executable cell I get the following error when calling `_, _ = pyntViz.plotIntensity(rawTimeSeries, hz, title='Cal590 Raw Traces')`

```
OSError: 'seaborn-dark-palette' is not a valid package style, path of style file, URL of style file, or library style name (library styles are listed in `style.available`)
```

I thought maybe installing seaborn could fix the issue but actually conda installing seaborn broke the pyneurotrace imports in the notebook.
