### 2024-06-18 6ca4045
Update README.md
GPU installation info
