### 2024-06-18 padster/pyNeuroTrace#4 @Peter-Hogg
There is a flag for the pip install command that will include Cupy (pip install pyNeuroTrace[GPU]) which will try to install Cupy with other dependencies. This is currently noted in the documentation but not the ReadMe. It is a bit tricky because Cupy is dependent on several requirements that must be set up for it to work correctly.

I agree with your suggestion and I have updated the ReadMe to detail installing pyNeuroTrace with Cupy. I'll add some of what I've written into the documentation as well so that it had additional detail and links to Cupy documentation as well.

I've also noticed that numpy 2.0 breaks several functions and  I will have to updated the requirements to use a numpy version <2.0 until I can investigate why they are breaking and update them accordingly.
