### 2024-06-18 a19b3ad
Numpy Bug Fix
