+++ pyproject.toml
@@ -5,13 +5,13 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyNeuroTrace"
-version = "0.9.84"
+version = "0.9.85"
 dependencies = [
 "ipython",
 "ipywidgets",
 "matplotlib",
 "matplotlib-scalebar",
-"numpy==1.24.4",
+"numpy==1.23.0",
 "pytest",
 "PyQt5",
 "scipy",
