+++ requirements.txt
@@ -1,3 +1,5 @@
+setuptools>=58.0.0
+wheel
 ipython
 ipywidgets
 matplotlib
