+++ requirements.txt
@@ -1,5 +1,6 @@
 setuptools>=58.0.0
 wheel
+distutils
 ipython
 ipywidgets
 matplotlib
