+++ requirements.txt
@@ -1,6 +1,5 @@
-setuptools>=58.0.0
+setuptools>=70.0.0
 wheel
-distutils
 ipython
 ipywidgets
 matplotlib
