### 2024-06-18 padster/pyNeuroTrace#5 @Peter-Hogg
Matplotlib depreciated that color palette apparently. I can still use it by changing the name to 'seaborn-v0_8-dark-palette'. I've updated the viz module with other bugs and made a new release fixing this issue.
