### 2024-06-27 #6877 @Peter-Hogg
> Hi @Peter-Hogg , could you please fix broken links to images in the [README.md](https://github.com/padster/pyNeuroTrace/blob/master/pyneurotrace/README.md) file?

I can do that
