### 2024-07-05 padster/pyNeuroTrace#4 @danielmk
pyneurotrace installs now without cupy so this is closed. If I run into problems installing with cupy I will open a separate issue.
