### 2024-07-05 padster/pyNeuroTrace#6 @danielmk
some functions missing docstring
Related to review: https://github.com/openjournals/joss-reviews/issues/6877

I'm working through the point "Functionality documentation".

- Functions in analysis.py don't have docstrings
- `nonNegativeDeconvolution` in nndFilter.py does not have a docstring
- in gpu, `ewma`, `cusum` and `thresholdEvents` don't have docstrings

For underscored functions I think it's ok if they are missing docstrings because they are not user facing.
