### 2024-07-05 padster/pyNeuroTrace#7 @danielmk
gpu functionality and performance
Related to https://github.com/openjournals/joss-reviews/issues/6877

I was able to install with cupy and I would like to confirm the GPU performance claim. Would it be possible to get the data that was used in the paper (Figure4)? And the code that was used? Thank you.
