### 2024-07-05 padster/pyNeuroTrace#7 @Peter-Hogg
Yes, I'll add a notebook to the repo soon. I'm currently traveling so it might be a day or two.
