### 2024-07-05 #6877 @stephenlenzi
Hi @Peter-Hogg in the example notebook `_, _ = plotIntensity(rawTimeSeries, hz, title='Cal590 Raw Traces')` fails because plotIntensity is not imported. Could you add the import?

Also, could you please either add jupyter as a dependency? Alternatively since it is not strictly needed for the software but only for the examples you could indicate in the README that the user needs to pip install jupyter to use the notebooks. I realise its obvious to most, but I suspect not everyone will realise this.
