### 2024-07-09 padster/pyNeuroTrace#9 @stephenlenzi
Comparison of the tool with other commonly used packages
Related to review: https://github.com/openjournals/joss-reviews/issues/6877

Personally I'd like to see a bit more detail about the state of the field and where this tool fits into the wider landscape. For example there are other softwares for dealing with calcium imaging data (e.g. Suite2P, CaImAn) and although they have a slightly different focus, they do seem to have overlapping functionality particularly in terms of event detection and estimating df/f. I think it is important to highlight how pyNeuroTrace compares and/or complements these and other existing pipelines.

On a (possibly) related note, the statement of need in paper.md is different to paper.pdf (is the paper.md more up to date?). The pdf doesn't mention other softwares explicitly at all, whereas the .md does highlight a couple of examples.
