### 2024-07-09 padster/pyNeuroTrace#10 @stephenlenzi
minor manuscript corrections
Related to review: https://github.com/openjournals/joss-reviews/issues/6877

Hi I think I found a couple of minor issues in the text of the manuscript:

- There is currently no reference in the text to figure 3, possibly it is a typo as Figure 2 is mentioned where I was expecting Figure 3 to be mentioned: "Figure 2 shows the average evoked response to a full-field OFF and ON stimuli from the data in Figure 1"
- There is what appears to be a broken link to Figure 3 in the legend: "Figure 3: Average stimulus-locked responses from the same in vivo imaging experiment of the dendritic branch in {fig:heatmap}"
- “To vectorise the functions several where modified” - I think there is a typo and "where" should be "were"
