### 2024-07-09 padster/pyNeuroTrace#9 @stephenlenzi
Just to follow up on the last point - I was reading the paper.pdf instead of the draft MS (assuming they were the same, but they're not.. my mistake). I see that the statement of need in the draft MS does indeed mention alternative softwares.
