### 2024-07-09 padster/pyNeuroTrace#10 @stephenlenzi
I may be mistaken here, but I think there is also a typo in the okada filter section:

"The value for 𝛼, which is a coefficient, should be selected so that the product of 𝑥𝑡 − 𝑥𝑡
and 𝑥𝑡 − 𝑥𝑡+1  causes a sufficiently steep sigmoid curve which functions a binary filter in the equation."

I think this should be:

"The value for 𝛼, which is a coefficient, should be selected so that the product of 𝑥𝑡 − 𝑥𝑡-1
and 𝑥𝑡 − 𝑥𝑡+1  causes a sufficiently steep sigmoid curve which functions a binary filter in the equation."
