### 2024-07-16 #6877 @danielmk
Review going well. Mainly waiting for code and data to confirm the GPU performance. See https://github.com/padster/pyNeuroTrace/issues/7
