### 2024-07-16 padster/pyNeuroTrace#9 @danielmk
The draft version I have (10.21105.joss.06877.pdf) references CaImAn but does not reference Suite2P. I think it would be important to reference Suite2P and explain the gap that the JOSS submission is filling in Suite2P functionality.
