### 2024-07-16 #6877 @stephenlenzi
Hi @adamltyson same here. All going well just awaiting the outcome of Daniel's issue so I can check off the performance and gpu-related items
