### 2024-07-16 #6877 @Peter-Hogg
Sorry for the delay I'm currently traveling. It's been a busier trip than I expected. I'm back home this weekend and will be able to push the extra notebooks and code to the repo.
