+++ paper/paper.md
@@ -13,6 +13,7 @@ authors:
     affiliation: 1
 
   - name: Peter William Hogg
+    corresponding: true 
     orcid: 0000-0003-2176-4977
     equal-contrib: true 
     affiliation: 1
@@ -70,7 +71,7 @@ $$
  x_{t} \leftarrow x_{t} + \frac{x_{t-1} + x_{t+1} - 2x_{t}}{2(1 + e^{-\alpha (x_{t} - x_{t-1})(x_{t} - x_{t+1})})}
 $$
 
-In this equation, $x_{t}$ is the value in the neural activity trace at time $t$. The value for $\alpha$, which is a coefficient, should be selected so that the product of $x_{t} - x_{t}$ and $x_{t} - x_{t+1}$ causes a sufficiently steep sigmoid curve which functions a binary filter in the equation. This function is equivalent to the following conditional states from Okada *et al.*:
+In this equation, $x_{t}$ is the value in the neural activity trace at time $t$. The value for $\alpha$, which is a coefficient, should be selected so that the product of $x_{t} - x_{t-1}$ and $x_{t} - x_{t+1}$ causes a sufficiently steep sigmoid curve which functions a binary filter in the equation. This function is equivalent to the following conditional states from Okada *et al.*:
 
 $$
  \text{If } (x_{t} - x_{t-1})(x_{t} - x_{t+1}) \leq 0 \
@@ -104,16 +105,16 @@ For individual or small numbers of activity traces, `pyNeuroTrace` has a line pl
 
 ![Six traces from neuron cell bodies imaged *in vivo* with a SLAP2 microscope. Tectal cells were bulk loaded with CAL-590-AM, a calcium senstive dye. Plotting individual traces better highlights the event shape \label{fig:traces}](docs/img/pynt_lineplot_cal590.png)
 
-Additionally, if a record of stimulus or trigger times is provided `pyNeuroTrace` can plot the average evoked response in a recording. \autoref{fig:traces} shows the average evoked response to a full-field OFF and ON stimuli from the data in \autoref{fig:heatmap}. This graph shows the evoked synaptic weights differ for the different visual stimuli presented to the animal. 
+Additionally, if a record of stimulus or trigger times is provided `pyNeuroTrace` can plot the average evoked response in a recording. \autoref{fig:AvgResponse} shows the average evoked response to a full-field OFF and ON stimuli from the data in \autoref{fig:heatmap}. This graph shows the evoked synaptic weights differ for the different visual stimuli presented to the animal. 
 
-![Average stimulus-locked responses from the same *in vivo* imaging experiment of the dendritic branch in {fig:heatmap} \label{fig:AvgResponse}. The response for each branch node is plotted for two visual stimuli, full field OFF and ON](docs/img/pyn_avg_response_on_off_9.png)
+![Average stimulus-locked responses from the same *in vivo* imaging experiment of the dendritic branch in \autoref{fig:heatmap} \label{fig:AvgResponse}. The response for each branch node is plotted for two visual stimuli, full field OFF and ON](docs/img/pyn_avg_response_on_off_9.png)
 
 # GPU Acceleration
[…]
