### 2024-07-19 padster/pyNeuroTrace#10 @Peter-Hogg
Thank you for reading the manuscript in detail. I have corrected all of the above typos.
