### 2024-07-19 padster/pyNeuroTrace#6 @Peter-Hogg
Thank you for closely inspecting the Python code. I've added the missing docstrings to the relevant functions.
