### 2024-07-19 padster/pyNeuroTrace#7 @Peter-Hogg
Sorry for the delay. I was hoping to be able to work on this during my trip, but my schedule was too busy. I've added the code used to generate the figures in the manuscript to the examples folder in a new Jupyter Notebook.
