### 2024-07-20 padster/pyNeuroTrace#9 @Peter-Hogg
I'll review Suite2P's current functionality. The last time I checked, it didn't calculate dF/F. If that is still the case, I'll add some comments to the manuscript about how this library fills that gap.

@stephenlenzi  I think the paper.pdf got out of sync with the paper.md because I forgot to regenerate the pdf version with Pandoc before pushing after editing the manuscript. I'll try not to do that again.
