### 2024-07-25 7a06abb
Manuscript updates
