+++ paper/paper.bib
@@ -156,3 +156,15 @@ doi = {10.1371/journal.pone.0157595}
   month = mar,
   pages = {583–591}
 }
+
+@article {Pachitariu2017,
+	author = {Pachitariu, Marius and Stringer, Carsen and Dipoppa, Mario and Schr{\"o}der, Sylvia and Rossi, L. Federico and Dalgleish, Henry and Carandini, Matteo and Harris, Kenneth D.},
+	title = {Suite2p: beyond 10,000 neurons with standard two-photon microscopy},
+	elocation-id = {061507},
+	year = {2017},
+	doi = {10.1101/061507},
+	publisher = {Cold Spring Harbor Laboratory},
+	URL = {https://www.biorxiv.org/content/early/2017/07/20/061507},
+	eprint = {https://www.biorxiv.org/content/early/2017/07/20/061507.full.pdf},
+	journal = {bioRxiv}
+}
