+++ paper/paper.md
@@ -42,7 +42,7 @@ Modern techniques for measuring neuronal activity using fluorescent biosensors a
 
 # Statement of need
 
-Many neuroscience labs using optophysiological methods, such as sampling neural activity using two-photon microscopy or fiber photometry, typically must create and constantly adjust functions and filters to analyze raw recordings. Currently, there is limited standardization for approaches for signal processing, with techniques and algorithms scattered throughout the literature such as [@Friedrich2017], requiring substantial framework overhead [@Giovannucci2019] or implemented in various programming languages other than Python [@Ortega2024]. Our library, `pyNeuroTrace`, seeks to address these problems by providing an analytic package written purely in Python specifically for neuronal activity data sets. Our package includes a collection of filters and algorithms relevant to optophysiological analysis that are implemented in a generalizable manner for time series data sets in either 1D arrays or a collection of recordings in 2D arrays. Additionally, with the increase in acquisition rates of new imaging techniques, we have implemented a subset of these algorithms using GPU-compatible code to significantly increase processing speeds to accommodate large datasets collected, such as those from large population sampling or ultra-fast kilohertz sampling rates. 
+Many neuroscience labs using optophysiological methods, such as sampling neural activity using two-photon microscopy or fiber photometry, typically must create and constantly adjust functions and filters to analyze raw recordings. Currently, there is limited standardization for approaches for signal processing, with techniques and algorithms scattered throughout the literature such as [@Friedrich2017], requiring substantial framework overhead [@Giovannucci2019] or implemented in various programming languages other than Python [@Ortega2024]. Additionally, other popular tools used to extract calcium signals from imaging experiments often have limited signal processing utilities, lacking functions for calculating $\Delta F/F$, forcing users to rely on independent methods [@Pachitariu2017]. Our library, `pyNeuroTrace`, seeks to address these problems by providing an analytic package written purely in Python specifically for neuronal activity data sets. Our package includes a collection of filters and algorithms relevant to optophysiological analysis that are implemented in a generalizable manner for time series data sets in either 1D arrays or a collection of recordings in 2D arrays. Additionally, with the increase in acquisition rates of new imaging techniques, we have implemented a subset of these algorithms using GPU-compatible code to significantly increase processing speeds to accommodate large datasets collected, such as those from large population sampling or ultra-fast kilohertz sampling rates. 
 
 
 # Signal Processing
