### 2024-07-25 padster/pyNeuroTrace#9 @Peter-Hogg
I have added a brief statement referencing how pyNeuroTrace fills a gap in Suite2P's functionality. I can expand this further in the manuscript is needed
