### 2024-08-02 padster/pyNeuroTrace#11 @danielmk
Data sources and animal ethics clarification
Related to https://github.com/openjournals/joss-reviews/issues/6877

I want to clarify the source of the data for animal ethics reasons.

Figure 1 mentions Haas et al. 2001 so I assume the data shown is part of that paper. I assume the ethical approval is handled in that paper. (The paper actually contains no ethics statement but I assume that is because the type of experiment did not require approval in that jurisdiction and that was handled internally).

Are the data in Figures 2-4 from the same paper? If so, I think a short "data from Haas et al. 2001." note would be good for clarification. If not, the source should be made clear. If the data has not been published elsewhere, the paper needs an ethics statement or the authors should briefly clarify here why these in vivo experiments did not require ethics approval.
