### 2024-08-02 #6877 @danielmk
Hello @adamltyson I just ran the GPU code so I can confirm the performance claim. That was my last major item. I opened one last issue regarding the source of the data and possible ethics approval. That should be easy to clarify and when that issue is closed I am through my checklist.
