### 2024-08-02 padster/pyNeuroTrace#11 @Peter-Hogg
Thank you for this revision. The citation for Hass et al. 2001 pertains to the labeling technique, which is not commonly used. The presented data in our manuscript are unpublished results from pilot studies conducted for other research projects.  I will add a heading and ethics statement to the manuscript. All work was approved by the University of British Columbia Animal Care Committee, protocols  A19-0297 and A24-0049, in full accordance with the rules and regulations of the Canadian Council of Animal Care, which is our national oversight organization.
