### 2024-08-02 #6877 @Peter-Hogg
I want to thank the reviewers for their insights and feedback. I've added a statement on ethics approval to the manuscript. If the remaining issues on the repository have been addressed to a satisfactory level I'll close them.
