### 2024-08-04 padster/pyNeuroTrace#11 @danielmk
Thank you for adding this clarification. If that statement is added that closes this issue for me.
