### 2024-08-05 #6877 @adamltyson
Thanks @Peter-Hogg. @danielmk your checklist is complete, so you can you confirm you're happy with the changes to the manuscript, and you're happy to recommend acceptance?

@stephenlenzi is there anything outstanding on your side?
