### 2024-08-05 #6877 @danielmk
@adamltyson Yes I recommend acceptance. I just want to note that I made the authors add a statement on animal ethics, since the in-vivo data in the paper is not published elsewhere. From my perspective the ethics statement is appropriate, but maybe JOSS has some specific guidelines that I am not aware of. Since you have more experience with JOSS, it would be great if you could have a quick look in the [paper.pdf](https://github.com/padster/pyNeuroTrace/blob/master/paper/paper.pdf) and confirm that the ethics statement is fine. Otherwise that's it from me and I am satisfied with the paper. lLt me know if I can help with anything else.
