### 2024-08-05 #6877 @stephenlenzi
Thanks @adamltyson, there is nothing left outstanding on my side. I also recommend acceptance.
