### 2024-08-05 #6877 @adamltyson
Hi @danielmk, @Peter-Hogg. If this data is original, then there are some [specific JOSS guidelines](https://joss.readthedocs.io/en/latest/policies.html). Thanks for adding the relevant approvals @Peter-Hogg, but would you also be able to report either the [ARRIVE](https://arriveguidelines.org/) or [PREPARE](https://doi.org/10.1177/0023677217724823) guidelines too?

Then it looks like we're ready to go. FYI I'm away for the rest of this week, so all being well I will start the acceptance process next week.

Thanks all!
