### 2024-08-06 #6877 @Peter-Hogg
@adamltyson I hope you enjoy your trip. I'll work on the ARRIVE documentation. Should I add a methods section to the manuscript detailing sample preparation as well? That's currently not included in the manuscript as I thought it was outside of the scope of paper when I submitted it. However, these reports from ARRIVE and PREPARE would like them included to make the work more reproducible.
