### 2024-08-12 #6877 @adamltyson
Yes, please add whatever is needed to adhere to the ARRIVE or PREPARE guidelines. It's of course not related to the software itself, but if this is original data, then it's JOSS policy that these details are added.
