+++ paper/paper.bib
@@ -168,3 +168,41 @@ doi = {10.1371/journal.pone.0157595}
 	eprint = {https://www.biorxiv.org/content/early/2017/07/20/061507.full.pdf},
 	journal = {bioRxiv}
 }
+
+@article{Hogg2021,
+  title = {Bulk Dye Loading for In Vivo Calcium Imaging of Visual Responses in Populations of Xenopus Tectal Neurons},
+  volume = {2022},
+  ISSN = {1559-6095},
+  url = {http://dx.doi.org/10.1101/pdb.prot106831},
+  DOI = {10.1101/pdb.prot106831},
+  number = {1},
+  journal = {Cold Spring Harbor Protocols},
+  publisher = {Cold Spring Harbor Laboratory},
+  author = {Hogg,  Peter W. and Haas,  Kurt},
+  year = {2021},
+  month = mar,
+  pages = {pdb.prot106831}
+}
+
+@article{Forman2017,
+  title = {Openspritzer: an open hardware pressure ejection system for reliably delivering picolitre volumes},
+  volume = {7},
+  ISSN = {2045-2322},
+  url = {http://dx.doi.org/10.1038/s41598-017-02301-2},
+  DOI = {10.1038/s41598-017-02301-2},
+  number = {1},
+  journal = {Scientific Reports},
+  publisher = {Springer Science and Business Media LLC},
+  author = {Forman,  C. J. and Tomes,  H. and Mbobo,  B. and Burman,  R. J. and Jacobs,  M. and Baden,  T and Raimondo,  J. V.},
+  year = {2017},
+  month = may 
+}
+
+@misc{mbfSLAP2,
+  author = {Podgorski, Kaspar},
+  title = {SLAP2 – Two Photon Microscope Kit},
+  year = {2021},
+  publisher = {MBF Bioscience},
+  journal = {Product Page},
+  url = {https://www.mbfbioscience.com/products/slap2}
+}
