+++ paper/paper.md
@@ -97,7 +97,7 @@ The event detection module uses several strategies to identify neuronal activity
 # Visualization
 `pyNeuroTrace` has several built-in visualization tools depending on the format of the data. 2D arrays of neuronal timeseries can be displayed as heat maps \autoref{fig:heatmap} or as individual traces \autoref{fig:traces}. The heatmap is a useful visualization tool for inspecting many traces at once; additionally, at the bottom of the plot, the stimuli timing is displayed if provided \autoref{fig:heatmap}. This functionality allows for quick visual inspection of  from a population of neurons or signals sampled across a neuronal structure, such as a dendritic arbor.
 
-One of the in-built visualizations is specific to the data structure generated by a custom acousto-optic random access two-photon microscope [@Sakaki2020] \autoref{fig:heatmap}. This microscope uses acousto-optic deflectors (AODs) to perform inertia-free scanning between preselected points of interest, allowing for extremely fast acquisition rates for sampling neuronal activity throughout complex 3D neural structures. The scan engine of the microscope allows for random-access sampling for imaging activity across the entire dendritic arbor morphology of a single neuron [@Sakaki2020]. This type of imaging does not generate a traditional image. The microscope instead links acquired neuronal traces to points of interest organized into a hierarchical tree structure representing the neuronal morphology in a complex data file. `pyNeuroTrace` will use additional information from this microscope to link structure and function by color-coding where neuronal activity comes from in a sampled neuron. 
+One of the in-built visualizations is specific to the data structure generated by a custom acousto-optic random access multi-photon (RAMP) microscope [@Sakaki2020] \autoref{fig:heatmap}. This microscope uses acousto-optic deflectors (AODs) to perform inertia-free scanning between preselected points of interest, allowing for extremely fast acquisition rates for sampling neuronal activity throughout complex 3D neural structures. The scan engine of the microscope allows for random-access sampling for imaging activity across the entire dendritic arbor morphology of a single neuron [@Sakaki2020]. This type of imaging does not generate a traditional image. The microscope instead links acquired neuronal traces to points of interest organized into a hierarchical tree structure representing the neuronal morphology in a complex data file. `pyNeuroTrace` will use additional information from this microscope to link structure and function by color-coding where neuronal activity comes from in a sampled neuron. 
 
 ![An example of a heatmap generated by pyNeuroTrace. A) An example of a lab-specific visualization of points of interest across a dendritic arbor imaged *in vivo* with an AOD random-access two-photon microscope. Plotted data was recorded from a single dendritic branch *in vivo* using a random access two-photon microscope. The indicator is a GCaMP6m, a plasmid encoding the GECI was single-cell electroporated [@Haas2001] B) Heatmap of visually evoked responses imaged at 148 Hz, showing intensity (yellow being highest), in addition to stim indicators along the X axis, and branch indicator along the Y. \label{fig:heatmap}](docs/img/pynt_Branch_Heatmap.png)
 
@@ -142,6 +142,22 @@ $$emwa = \frac{c[i]}{n[i]}$$
 ![Overlay of the EWMA calculations using the CPU implementation and GPU approximation in red and blue. The difference in values from the output is also plotted. The EWMA traces diverge after the initial window length of the GPU version's kernel, but never by more than a factor of 1e-11. \label{fig:ewmaCPUvsGPU}](docs/img/ewma_CPUvsGPU.png)
 
[…]
