+++ paper/paper.md
@@ -155,7 +155,7 @@ Single-cell electroporation (SCE) was used to label individual cells with fluore
 CAL-590-AM (AAT Bioquest) was pressure-injected using a modified protocol previously described[@Hogg2021]. A 10 mM CAL-590-AM stock solution was prepared in DMSO with 20% Pluronic Acid (ThermoFisher). This solution was diluted 1:10 in Calcium Free-Ringers as previously described for Oregon Green™ 488 BAPTA-1 imaging protocol[@Hogg2021]. The tadpole was anesthetized using 0.02% MS-222 in Steinberg's solution before injection. A glass pipette filled with the dye was inserted into the optic tectum. The dye was pressure-injected using 0.02 MPa $N_2$ gas and an OpenSpritzer MicroInjector (NeuroGig)[@Forman2017]. 
 
 ## *In vivo* Imaging of Neuronal Activity
-Neuronal activity recordings were performed using two-photon microscopy. All animals were reversibly paralyzed by bath application of 2 mM pancuronium dibromide for 5 minutes (Tocris) before being transferred to a custom imaging chamber [@Sakaki2020]. The GCaMP6m expressing neuron was imaged using a custom AOD-based RAMP microscope previously described [@Sakaki2020]. A Ti:Sapphire laser (Chameleon Vision II, Coherent) tuned to 920 nm was used for two-photon excitation, and emitted light was captured using a 60x, 1.1 NA water-immersion objective (LUMPL, Olympus). Tectal cells filled with CAL-590 AM were imaged using a custom microscope built from a commercial kit [@mbfSLAP2]. The excitation source was a 1030 nm custom Ytterbium-doped fiber laser with an average power of 68 W, 150 fs pulse width, and 40 MHz repetition rate (Tangerine-SP, Amplitude). Full-field raster imaging was collected at 13.5 Hz through a 25x 1.0 NA water-immersion objective (HC IRAPO, Leica) \autoref{fig:traces}. The microscope's integrated scan mode was used to capture calcium signals at 2 kHz from a branch of a neuron labeled with CAL-590 3k Dextran \autoref{fig:CPUvsGPU}. All signal processing was performed using pyNeuroTrace.
+Neuronal activity recordings were performed using two-photon microscopy. All animals were reversibly paralyzed by bath application of 2 mM pancuronium dibromide for 5 minutes (Tocris) before being transferred to a custom imaging chamber [@Sakaki2020]. The GCaMP6m expressing neuron was imaged using a custom AOD-based RAMP microscope previously described [@Sakaki2020]. A Ti:Sapphire laser (Chameleon Vision II, Coherent) tuned to 910 nm was used for two-photon excitation, and emitted light was captured using a 60x, 1.1 NA water-immersion objective (LUMPL, Olympus). Tectal cells filled with CAL-590 AM were imaged using a custom microscope built from a commercial kit [@mbfSLAP2]. The excitation source was a 1030 nm custom Ytterbium-doped fiber laser with an average power of 68 W, 150 fs pulse width, and 40 MHz repetition rate (Tangerine-SP, Amplitude). Full-field raster imaging was collected at 13.5 Hz through a 25x 1.0 NA water-immersion objective (HC IRAPO, Leica) \autoref{fig:traces}. The microscope's integrated scan mode was used to capture calcium signals at 2 kHz from a branch of a neuron labeled with CAL-590 3k Dextran \autoref{fig:CPUvsGPU}. All signal processing was performed using pyNeuroTrace.
 
 
 
