### 2024-08-13 #6877 @adamltyson
Thanks @Peter-Hogg. I will give the paper itself a once-over. In the meantime, would you be able to go through the list of "Additional Author Tasks" above?

Could you also double check the missing DOIs too, i.e. check if a DOI can be added? I think they're probably OK because I seem to remember that NeurIPS doesn't give DOIs, and one is a commercial product.
