### 2024-08-13 padster/pyNeuroTrace#12 @adamltyson
Make the link to the documentation easier to find.
The documentation is key for anyone new to the software, and I think the only link is the badge? It would be helpful to make this link really obvious, maybe by:
- Adding it as the repo URL in the sidebar
- Adding it in the README text

Apologies if I missed it somewhere.
