### 2024-08-13 fcbf25e
Define GECI abbreviation and add some missing spaces
