+++ paper/paper.md
@@ -92,7 +92,7 @@ Essentially, in each trace the Okada filter replaces the point $x_{t}$ with the 
 When the kinetics of biosensors are known, nonnegative deconvolution (NND) can be useful to deconvolve raw or $\Delta F/F$ traces back into the most likely discrete triggers of increases in the sensors. `pyNeuroTrace` includes an implementation of an algorithm that can efficiently compute NND in linear time [@Podgorski2012]. This can also be useful for denoising, as well particularly good at finding smaller magnitude events in fluorescent imaging that are often obfuscated by machine noise [@Podgorski2012].
 
 # Event Detection
-The event detection module uses several strategies to identify neuronal activity events in time series datasets. These methodologies have been previously discussed and compared by Sakaki *et al.* [@Sakaki2018]. These include two generalizable methods and one that requires prior knowledge of recorded event shape. The generalizable methods include filtering the signal through an exponentially weighted moving average (ewma) or cumulative sum of movement above the mean (cusum). The final filter is a matched filter that finds the probability of the trace matching a previously defined shape, such as one described by an exponential rise and decay of calcium signal generated by a GECI. Computationally efficient algorithms were selected, allowing for the possibility of applying them in real-time during experiments.
+The event detection module uses several strategies to identify neuronal activity events in time series datasets. These methodologies have been previously discussed and compared by Sakaki *et al.* [@Sakaki2018]. These include two generalizable methods and one that requires prior knowledge of recorded event shape. The generalizable methods include filtering the signal through an exponentially weighted moving average (ewma) or cumulative sum of movement above the mean (cusum). The final filter is a matched filter that finds the probability of the trace matching a previously defined shape, such as one described by an exponential rise and decay of calcium signal generated by a genetically encoded calcium indicator (GECI). Computationally efficient algorithms were selected, allowing for the possibility of applying them in real-time during experiments.
 
 # Visualization
 `pyNeuroTrace` has several built-in visualization tools depending on the format of the data. 2D arrays of neuronal timeseries can be displayed as heat maps \autoref{fig:heatmap} or as individual traces \autoref{fig:traces}. The heatmap is a useful visualization tool for inspecting many traces at once; additionally, at the bottom of the plot, the stimuli timing is displayed if provided \autoref{fig:heatmap}. This functionality allows for quick visual inspection of  from a population of neurons or signals sampled across a neuronal structure, such as a dendritic arbor.
@@ -148,11 +148,11 @@ To demonstrate the differences between the CPU and GPU implementations of the EW
 Free-swimming albino *Xenopus laevis* tadpoles were reared in 10% Steinberg's solution and kept in a 12-hour light/dark cycle at 22&deg;C. All experiments were performed on Stage 48 tadpoles. 
 
 ## *In vivo*  Single-Cell Electroporation of DNA and Dye
[…]
