### 2024-08-13 #6877 @adamltyson
The paper looks good @Peter-Hogg. I've raised [a PR with some tiny changes](https://github.com/padster/pyNeuroTrace/pull/13) then I think we're ready to go.

I also raised [an issue](https://github.com/padster/pyNeuroTrace/issues/12) about making the docs more obvious. It's not required for acceptance, but I think it would be very helpful for users!
