### 2024-08-13 padster/pyNeuroTrace#12 @padster
Good suggestions -

Documentation link added to sidebar:
![sidebar](https://github.com/user-attachments/assets/d80db7f5-46a2-4242-8214-b48392dfca18)

And README:
![readme](https://github.com/user-attachments/assets/0a0312df-b3f3-43f0-bc13-80bca9c440c6)
