+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyNeuroTrace"
-version = "0.9.86"
+version = "1.0.0"
 dependencies = [
 "setuptools>=70.0.0",
 "ipython",
