+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyNeuroTrace"
-version = "1.0.0"
+version = "1.0.1"
 dependencies = [
 "setuptools>=70.0.0",
 "ipython",
