### 2024-08-15 #6877 @Kevin-Mattheus-Moerman
@Peter-Hogg as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however the below is a final point that requires your attention:

- [ ] `senstive` should read `sensitive`
