### 2024-08-15 #6877 @Peter-Hogg
@Kevin-Mattheus-Moerman, thanks for catching the typo; I've edited the paper with the most recent push. I've also removed the default license from the archive. Thanks everyone who has helped with the review process.

Cheers,

Peter
