### 2024-08-15 39c4a4b
Create CITATION.cff
