### 2024-08-16 #6877 @Kevin-Mattheus-Moerman
@Peter-Hogg et al. congratulations on this JOSS paper!

@adamltyson thanks for editing!

And a special thanks to the reviewers: @stephenlenzi, @danielmk !!
